src/scrapeyahoo/pyproject.toml
[tool.poetry]
name = "scrapeyahoo"
version = "0.1.0"
description = ""
authors = ["Aditya Pillai <github@pillai.io>"]

[tool.poetry.dependencies]
python = "^3.8"
yfinance = "^0.1.54"
gitpython = "^3.1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
